# src/main/resources/static/docs/openapi.yaml
openapi: 3.0.1
info:
  title: Gathering API
  description: Gathering 서비스 REST API 문서 (REST Docs 기반)
  version: 1.0.0
servers:
- url: http://localhost
tags: []
paths:
  /login:
    post:
      tags:
      - 인증
      - login
      summary: 로그인
      description: 사용자 로그인 API - JWT 토큰을 쿠키로 반환
      operationId: auth-login로그인
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/login-733236850"
            examples:
              로그인:
                value: "{\"email\":\"test@example.com\",\"password\":\"qARwtu22gWm4DtnttjuQbg==\"\
                  }"
              auth-login:
                value: "{\"email\":\"test@example.com\",\"password\":\"qARwtu22gWm4DtnttjuQbg==\"\
                  }"
      responses:
        "200":
          description: "200"
  /users/join:
    post:
      tags:
      - 사용자
      - users
      summary: 회원가입
      description: 새로운 사용자를 등록합니다
      operationId: users-join사용자-가입
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/users-join-1381348445"
            examples:
              사용자-가입:
                value: "{\"email\":\"test@example.com\",\"password\":\"qARwtu22gWm4DtnttjuQbg==\"\
                  ,\"nickname\":\"테스터\",\"name\":\"홍길동\",\"phoneNumber\":\"01012345678\"\
                  }"
              users-join:
                value: "{\"email\":\"test@example.com\",\"password\":\"qARwtu22gWm4DtnttjuQbg==\"\
                  ,\"nickname\":\"테스터\",\"name\":\"홍길동\",\"phoneNumber\":\"01012345678\"\
                  }"
      responses:
        "204":
          description: "204"
  /users/me:
    get:
      tags:
      - users
      operationId: users-me
      responses:
        "401":
          description: "401"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/users-tsid-1331249008"
              examples:
                users-me-no-token:
                  value: "{\"code\":\"INVALID_TOKEN\",\"message\":\"토큰이 유효하지 않습니다\"\
                    }"
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/users-me810629779"
              examples:
                users-me:
                  value: "{\"tsid\":\"1234567890123\"}"
  /users/{tsid}:
    get:
      tags:
      - users
      - 사용자
      summary: 정지된 사용자 조회
      description: 정지된 사용자를 조회할 경우 404 에러를 반환합니다
      operationId: users-getusers-get-bannedusers-get-deleted사용자-조회사용자-조회-삭제됨사용자-조회-정지됨
      parameters:
      - name: tsid
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "404":
          description: "404"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/users-tsid-1331249008"
              examples:
                users-get-deleted:
                  value: "{\"code\":\"USER_DELETED\",\"message\":\"삭제된 사용자입니다.\"}"
                사용자-조회-정지됨:
                  value: "{\"code\":\"USER_BANNED\",\"message\":\"사용이 제한된 사용자입니다.\"\
                    }"
                users-get-banned:
                  value: "{\"code\":\"USER_BANNED\",\"message\":\"사용이 제한된 사용자입니다.\"\
                    }"
                사용자-조회-삭제됨:
                  value: "{\"code\":\"USER_DELETED\",\"message\":\"삭제된 사용자입니다.\"}"
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/users-tsid-982397607"
              examples:
                users-get:
                  value: "{\"nickname\":\"테스터\",\"name\":\"홍길동\",\"profileImageUrl\"\
                    :\"https://example.com/profile.jpg\"}"
                사용자-조회:
                  value: "{\"nickname\":\"테스터\",\"name\":\"홍길동\",\"profileImageUrl\"\
                    :\"https://example.com/profile.jpg\"}"
components:
  schemas:
    users-me810629779:
      required:
      - tsid
      type: object
      properties:
        tsid:
          type: string
          description: 사용자 고유 ID
    login-733236850:
      required:
      - email
      - password
      type: object
      properties:
        password:
          type: string
          description: AES 암호화된 비밀번호 (Base64 인코딩)
        email:
          type: string
          description: 이메일 주소
    users-tsid-1331249008:
      required:
      - code
      - message
      type: object
      properties:
        code:
          type: string
          description: 에러 코드
        message:
          type: string
          description: 에러 메시지
    users-tsid-982397607:
      required:
      - name
      type: object
      properties:
        name:
          type: string
          description: 사용자 이름
        nickname:
          type: string
          description: 닉네임
          nullable: true
        profileImageUrl:
          type: string
          description: 프로필 이미지 URL
          nullable: true
    users-join-1381348445:
      required:
      - email
      - name
      - password
      - phoneNumber
      type: object
      properties:
        password:
          type: string
          description: AES 암호화된 비밀번호 (Base64 인코딩)
        phoneNumber:
          type: string
          description: "전화번호 (숫자만 입력, 예: 01012345678)"
        name:
          type: string
          description: 사용자 이름
        nickname:
          type: string
          description: 닉네임 (선택 사항)
          nullable: true
        email:
          type: string
          description: 이메일 주소 (이메일 형식)